+++ change/@fluentui-react-internal-2021-02-01-17-12-54-feat-react-theme-provider.json
@@ -1,0 +1,8 @@
+{
+  "type": "prerelease",
+  "comment": "move FocusZone to react-internal",
+  "packageName": "@fluentui/react-internal",
+  "email": "olfedias@microsoft.com",
+  "dependentChangeType": "patch",
+  "date": "2021-02-01T16:12:51.690Z"
+}
