+++ packages/react-focus/package.json
@@ -45,12 +45,8 @@
     "react-test-renderer": "^16.3.0"
   },
   "dependencies": {
-    "@fluentui/keyboard-key": "^0.2.13",
-    "@fluentui/merge-styles": "^8.0.0-beta.4",
-    "@fluentui/react-theme-provider": "^1.0.0-beta.18",
+    "@fluentui/react-internal": "^8.0.0-beta.40",
     "@fluentui/set-version": "^8.0.0-beta.1",
-    "@fluentui/style-utilities": "^8.0.0-beta.13",
-    "@fluentui/utilities": "^8.0.0-beta.10",
     "tslib": "^1.10.0"
   },
   "peerDependencies": {
