+++ packages/react-internal/src/FocusZone.ts
@@ -1,1 +1,1 @@
-export * from '@fluentui/react-focus';
+export * from './components/FocusZone';
